Add helper methods for iterating over characters
A common need among these functions is iterating over characters within a word. This commit adds two helper methods
to make this easier

  * _forEachCharacter(string, function)
    Call the callback function with each character in a string.

  * _forEachCharacterInWord(string, function)
    Call the callback function for each character in each word (i.e., spaces are not passed down)

Also added: _ifNullOrUndefined is a simple helper to check if the passed value is null or undefined

Refactors the code in caesar.js to use the new helpers.